Save _hashable_content tuple to be computed once from constructor,
That is this tuple itself cached too.
Not that it is half-hearted decision. The situation is possible,
when the hashes of arguments can be changed themselves, but there
is no guarantee that the order of arguments will be appropriated
them.

Anather disadvantage, that a new __slot__ for Basic class is added.

But this commit reduse the number of `sorted` calls:

This commit:
271902/264242    1.228    0.000    6.916    0.000 {sorted}

Previouse one:
556327/548246    2.156    0.000    8.342    0.000 {sorted}

Master (when sorting of argument is an Add class constructor):
175964/169169    0.705    0.000    5.757    0.000 {sorted}