Remove random sorting of Add arguments, insert nested Add inplace.
While cycling in Add.flatten the recursion is organized as
a cycle, which in particular two problems solves.
But with reordering.

1. The task to collect similar terms

    2*x**2 + y + 5*x**2   --> 7*x**2 + y

was solved originaly with the help of the dictionary
(`{x**2:7,y:S.One}`) but firther cycle by keys reorder them
randomly (or by hash).

In this commit The way of accumulation is changed, so do not break
original order.

2. The task of nested Adds was solved with the help of the recursion
which is organized as a cycle.

    Add(a, Add(b, c), d) --> Add(a, d, b, c)

In this commit nested Add's are inserted in-place now:

    Add(a, Add(b, c), d) --> Add(a, b, c, d)