Introduce PowerE series (former TaylorSeries).
The former TaylorSeries are renamed to PowerESeries as it becames
something else. PowerE series is the same as Power series but only with
coefficients devided by n! Thes is analog of Exponentional Generating
Function.
The algebra is the same, ony the convinience of calculations.

The TalorSeries will be the extand PowerSeries but in difference point.
Note also, that the main mission of TalorSeries is the obtain series expanyop from
function.
Main mission of Generationg Functions is obtain the function from sequence.

And Formal Power series provide the formal mechanizm with the operations.

while Power Series can solve the tasks of convergence or calculations of series.