Fix `cacheit` wrapper to cache `Add(x,y)` and `Add(y,x)` correctly.
Before this commite the Add(x,y) and Add(y,x) yield different objects.

    >>> Add(x,y) is Add(y,x)
    False

It was because `cacheit` is called before `__new__` method.
In turn, the flatten method is calls inside the __new__.

In this pull request we inject the simple flatting procedure to the cacheit wrapper.

Notes:

Only one tests became to be a failure: "test_heurisch_symbolic_coeffs" in "test_risch"
It is marked as XFAIL now.